Add mkbootimg package
The mkbootimg utility will be used to build the kernel boot images
for the projects whose bootloaders expect Android-style boot images.

Previously it was expected that a pre-built mkbootimg is provided in
the root of the kernel source tree. But that lets to build the
OpenELEC for such projects only on the platform where the mkbootimg
was built.

Closes #11